fix(surface-app): `.` must import nothing — a Vite config loads it under Node
`ci::dev-smoke` went red on the previous commit and it was right: a Vite config
is loaded by NODE's ESM loader, not by a bundler, and Node cannot resolve this
package's extensionless relative imports. kolu's own `packages/client/vite.config.ts`
imports `@kolu/surface-app` for `ASSET_DIR` and `NOTIFICATION_SW_SOURCE`, so the
re-export `index.ts` grew during the lens pass (`export { … } from "./shellHead"`)
was `ERR_MODULE_NOT_FOUND` at every consumer's `vite dev`. `./vite`'s header has
always stated this constraint for itself; nobody had noticed it binds the package
ROOT just as hard.

- `shellHead.ts` is gone. `index.ts` owns the whole head prelude again — the
  commit family, the `<head>` locator, the modulepreload tags, and the one
  ordered splice — and imports NOTHING, which is now stated in its header and
  pinned by `nodeEsm.test.ts`: two spawns of Node's own loader, one per entry a
  consumer's Node loads directly (`.` and `/vite`). They fail in milliseconds, in
  the package that owns the mistake, where `ci::dev-smoke` takes three minutes
  and reads as a timeout waiting for localhost.
- `injectShellHead(html, { preloadHrefs, commit })` is therefore exported, and
  `SurfaceClientBuildResult` gains `preloadHrefs` so the thing that feeds it is
  published too — lowy's F1-A, which the Node constraint has now chosen for us.
  Both are in the Reference table.
- `modulePreload.ts` keeps the walk. The tag moves to `index.ts` beside the
  splice that writes it; what kept the two together — that `rel="modulepreload"`
  asserts the target is a module — is now enforced at the walk, which REFUSES a
  non-JS static import rather than filtering it, so every href reaching the tag
  has already earned the assertion.

The public API delta for this PR is now one export plus one result field, both
additive; drishti still needs no source change.
