+++ website/src/content/surface/ref-surface-app.mdx
@@ -194,8 +194,10 @@ siblings and the code splitting every consumer used to bolt on afterwards.
 | `surfaceApp()` | `/vite` | Vite plugin that injects the build commit onto the shell |
 | `buildSurfaceClient(opts)` | `/bun` | the Bun build: hashed assets, rewritten shell, siblings, splitting |
 | `SurfaceClientBuildOptions` | `/bun` | its options — entrypoint, `distDir`, the HTML template + placeholders, `extraAssets`, `publicDir`, `minify`, `commit` |
-| `SurfaceClientBuildResult` | `/bun` | `{ jsHref, assetHrefs, assets }` — the hashed URLs your **placeholders** were rewritten to (the JS entry, and one per extra asset), plus an `AssetReport` per **compressible** asset in the hashed dir (sourcemaps and already-compressed media are skipped, and absent from the report). The shell also carries a `modulepreload` link per static chunk; those come from the build graph, not from a placeholder, and are not reported |
+| `SurfaceClientBuildResult` | `/bun` | `{ jsHref, assetHrefs, preloadHrefs, assets }` — every hashed URL the shell now names (the JS entry, one per extra asset, and the entry's static chunks in load order), plus an `AssetReport` per **compressible** asset in the hashed dir (sourcemaps and already-compressed media are skipped, and absent from the report) |
 | `AssetReport` | `/bun` | `{ file, bytes, siblings }` — identity size and each sibling's size, keyed by `Content-Encoding` |
+| `injectShellHead(html, { preloadHrefs, commit })` | `.` | the whole head prelude in one splice — preload links first, then the build identity — for a caller templating its own shell from a build result |
+| `injectShellCommit(html, commit)` | `.` | that same splice with nothing to preload: the build identity alone |
 | `PRECOMPRESSED_ENCODINGS` | `.` | the `[encoding, suffix]` table BOTH halves read: `freshStaticLayer` to negotiate, the builder to emit |
 
 ```ts
@@ -228,7 +230,9 @@ for (const a of assets) console.log(a.file, a.bytes, a.siblings);
   and names those files in the shell, transitively, so the fetches start while
   the HTML is still parsing. A `dynamic-import` chunk is never preloaded —
   that would fetch on first paint exactly what `import()` was written to defer —
-  and an entry that split into nothing gets no tags at all.
+  and an entry that split into nothing gets no tags at all. A static import that
+  is not a JS module is **refused**, loudly: the bundler emitted a file this
+  build rewrites no placeholder for, so the shell would name it nowhere.
 </Aside>
 
 <Aside type="caution" title="The hashed dir belongs to the build">
